requirements.txt
dotenv==0.9.9
matplotlib==3.9.4
narwhals==2.11.0
numpy==2.0.2
pandas==2.3.3
pillow==11.3.0
pyarrow==21.0.0
pydeck==0.9.1
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
requests==2.32.5
streamlit==1.50.0

